Added env "FS_DISABLE_ENV" if set to true we do not touch anything in the config files

+++ run.sh
@@ -30,50 +30,56 @@ DEFAULT_FS_MOD_ENABLE_AV="true"
 DEFAULT_FS_MOD_ADD_JSON_CDR="true"
 DEFAULT_FS_SQLITE_MEMORY="false"
 
-rm -rf /etc/freeswitch/envVars/*
-for var in ${!DEFAULT_FS*}; do
-  t=${var/DEFAULT_/}
-  if [ -z ${!t} ]; then
-    echo "Using default for ${t}:${!var}"
-    eval ${t}=${!var}
-    export "${t}"
-  else
-    echo "Using override value for ${t}"
-  fi
-done
+echo "disable env: ${FS_DISABLE_ENV}"
 
+if [[ "${FS_DISABLE_ENV}" != "true" ]]; then
 
-echo -e "[default]\npassword=${FS_EVENT_PASSWORD}\nport=${FS_EVENT_PORT}\n" > /etc/fs_cli.conf
-
-
-for var in ${!FS_*}; do
-  if [[ $var == FS_MOD_ENABLE_* ]]; then
-    var1=${var#FS_MOD_ENABLE_*}
-    var2=${var1,,}
-    var3="mod_${var2}"
-    if [[ "${!var}" == "true" ]]; then
-      sed -i -e "s/.*<load module=\"${var3}\"\/>.*/<load module=\"${var3}\"\/>/g" /etc/freeswitch/autoload_configs/modules.conf.xml
-    else 
-      sed -i -e "s/.*<load module=\"${var3}\"\/>.*/<!--<load module=\"${var3}\"\/>-->/g" /etc/freeswitch/autoload_configs/modules.conf.xml
+  rm -rf /etc/freeswitch/envVars/*
+  for var in ${!DEFAULT_FS*}; do
+    t=${var/DEFAULT_/}
+    if [ -z ${!t} ]; then
+      echo "Using default for ${t}:${!var}"
+      eval ${t}=${!var}
+      export "${t}"
+    else
+      echo "Using override value for ${t}"
     fi
-  elif [[ $var == FS_MOD_ADD_* ]]; then
-    var1=${var#FS_MOD_ADD_*}
-    var2=${var1,,}
-    var3="mod_${var2}"
-    if [[ "${!var}" == "true" ]]; then
-      sed -i -e "s/.*<\/modules>.*/<load module=\"${var3}\"\/>\n<\/modules>/g" /etc/freeswitch/autoload_configs/modules.conf.xml
+  done
+
+
+  echo -e "[default]\npassword=${FS_EVENT_PASSWORD}\nport=${FS_EVENT_PORT}\n" > /etc/fs_cli.conf
+
+
+  for var in ${!FS_*}; do
+    if [[ $var == FS_MOD_ENABLE_* ]]; then
+      var1=${var#FS_MOD_ENABLE_*}
+      var2=${var1,,}
+      var3="mod_${var2}"
+      if [[ "${!var}" == "true" ]]; then
+        sed -i -e "s/.*<load module=\"${var3}\"\/>.*/<load module=\"${var3}\"\/>/g" /etc/freeswitch/autoload_configs/modules.conf.xml
+      else 
+        sed -i -e "s/.*<load module=\"${var3}\"\/>.*/<!--<load module=\"${var3}\"\/>-->/g" /etc/freeswitch/autoload_configs/modules.conf.xml
+      fi
+    elif [[ $var == FS_MOD_ADD_* ]]; then
+      var1=${var#FS_MOD_ADD_*}
+      var2=${var1,,}
+      var3="mod_${var2}"
+      if [[ "${!var}" == "true" ]]; then
+        sed -i -e "s/.*<\/modules>.*/<load module=\"${var3}\"\/>\n<\/modules>/g" /etc/freeswitch/autoload_configs/modules.conf.xml
+      fi
+    elif [[ $var == FS_* ]]; then
+      echo "<X-PRE-PROCESS cmd=\"set\" data=\"${var}=${!var}\"/>" > "/etc/freeswitch/envVars/${var}.xml"
     fi
-  elif [[ $var == FS_* ]]; then
-    echo "<X-PRE-PROCESS cmd=\"set\" data=\"${var}=${!var}\"/>" > "/etc/freeswitch/envVars/${var}.xml"
+  done
+
+  if [[ $FS_SIP_CAPTURE_SERVER != "false" ]]; then
+    sed -i -e 's/.*<param name="capture-server".*/<param name="capture-server" value="$${FS_SIP_CAPTURE_SERVER}"\/>/g' /etc/freeswitch/autoload_configs/sofia.conf.xml
   fi
-done
 
-if [[ $FS_SIP_CAPTURE_SERVER != "false" ]]; then
-  sed -i -e 's/.*<param name="capture-server".*/<param name="capture-server" value="$${FS_SIP_CAPTURE_SERVER}"\/>/g' /etc/freeswitch/autoload_configs/sofia.conf.xml
+  if [[ $FS_SQLITE_MEMORY != "false" ]]; then
+    sed -i -e 's/.*<!-- <param name="core-db-dsn" value="dsn:username:password" \/> -->.*/<param name="core-db-dsn" value="sqlite:\/\/memory:\/\/core.db" \/>/g' /etc/freeswitch/autoload_configs/switch.conf.xml
+  fi
+else
+  echo "Disabled Freeswitch Environment setup, using config files only"
 fi
-
-if [[ $FS_SQLITE_MEMORY != "false" ]]; then
-  sed -i -e 's/.*<!-- <param name="core-db-dsn" value="dsn:username:password" \/> -->.*/<param name="core-db-dsn" value="sqlite:\/\/memory:\/\/core.db" \/>/g' /etc/freeswitch/autoload_configs/switch.conf.xml
-fi
-
 exec "${@}"
